+++ FileMe.DAL/Classes/Folder.cs
@@ -4,7 +4,7 @@ namespace FileMe.DAL.Classes
 {
     public class Folder
     {
-        public virtual int Id { get; set; }
+        public virtual long Id { get; set; }
 
         public virtual string Name { get; set; }
